+++ build_windows.ps1
@@ -3,7 +3,7 @@ param(
     [string]$PythonPath = "python",
 
     [ValidatePattern("^\d+\.\d+\.\d+$")]
-    [string]$Version = "0.1.6",
+    [string]$Version = "0.1.7",
 
     [switch]$SkipValidation,
     [switch]$UseCurrentEnvironment,